Add Utility Functions and Validation Logic
